fix up removal of streamed content
git-svn-id: http://svn.vuze.com/public/client@19658 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus3/src/com/aelitis/azureus/core/devices/impl/TranscodeJobImpl.java
@@ -448,7 +448,7 @@ TranscodeJobImpl
 			state = ST_REMOVED;
 		}
 		
-		if ( delete_file ){
+		if ( delete_file && !isStream()){
 			
 			try{
 				transcode_file.delete( true );
